feat: Component development and folder structure setting
