+++ components/Footer.tsx
@@ -1,0 +1,11 @@
+/* eslint-disable prettier/prettier */
+import React from 'react';
+import {TouchableOpacity} from 'react-native';
+
+interface IProps {}
+
+const Footer: React.FC<IProps> = () => {
+  return <TouchableOpacity></TouchableOpacity>;
+};
+
+export default Footer;
