+++ utils/color/TextColor.ts
@@ -1,0 +1,6 @@
+/* eslint-disable prettier/prettier */
+export const TEXT_TITLE = '#171B21';
+export const TEXT_DDEFAULT = '#2E3440';
+export const TEXT_UNSELECTED = '#626978';
+export const TEXT_CAPTION = '#818995';
+export const TEXT_DISABLE = '#818995';
