+++ utils/color/UiColor.ts
@@ -1,0 +1,2 @@
+/* eslint-disable prettier/prettier */
+export const UI_WHITE = '#FFFFFF';
